Add a directory getting_started to ja.
