+++ website/i18n/ja/docusaurus-plugin-content-docs/current/getting_started/dart_pubspec/index.tsx
@@ -1,0 +1,9 @@
+import raw from "!!raw-loader!./raw.yaml";
+import codegen from "!!raw-loader!./codegen.yaml";
+
+export default {
+  raw,
+  hooks: raw,
+  codegen,
+  hooksCodegen: codegen,
+};
