+++ website/i18n/ja/docusaurus-plugin-content-docs/current/getting_started/dart_pubspec/raw.yaml
@@ -1,0 +1,6 @@
+name: my_app_name
+environment:
+  sdk: ">=2.17.0 <3.0.0"
+
+dependencies:
+  riverpod: ^2.1.3
