+++ website/i18n/ja/docusaurus-plugin-content-docs/current/getting_started/hello_world/index.tsx
@@ -1,0 +1,9 @@
+import raw from "!!raw-loader!./raw.dart";
+import codegen from "!!raw-loader!./main.dart";
+
+export default {
+  raw,
+  hooks: raw,
+  codegen,
+  hooksCodegen: codegen,
+};
