+++ website/i18n/ja/docusaurus-plugin-content-docs/current/getting_started/pubspec/hooks.yaml
@@ -1,0 +1,10 @@
+name: my_app_name
+environment:
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  flutter_hooks: ^0.18.0
+  hooks_riverpod: ^2.1.3
