+++ website/i18n/ja/docusaurus-plugin-content-docs/current/getting_started/pubspec/hooks_codegen.yaml
@@ -1,0 +1,15 @@
+name: my_app_name
+environment:
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  flutter_hooks: ^0.18.0
+  hooks_riverpod: ^2.1.3
+  riverpod_annotation: ^1.1.1
+
+dev_dependencies:
+  build_runner:
+  riverpod_generator: ^1.1.1
