+++ website/i18n/ja/docusaurus-plugin-content-docs/current/getting_started/pubspec/index.tsx
@@ -1,0 +1,11 @@
+import raw from "!!raw-loader!./raw.yaml";
+import hooks from "!!raw-loader!./hooks.yaml";
+import codegen from "!!raw-loader!./codegen.yaml";
+import hooksCodegen from "!!raw-loader!./hooks_codegen.yaml";
+
+export default {
+  raw,
+  hooks,
+  codegen,
+  hooksCodegen,
+};
